requirements.txt
alembic==1.13.1
bcrypt==4.1.2
blinker==1.7.0
click==8.1.7
colorama==0.4.6
dnspython==2.6.1
email-validator==2.1.0.post1
Flask==3.0.2
Flask-Bcrypt==1.0.1
Flask-Login==0.6.3
Flask-Migrate==4.0.5
Flask-Seeder==1.2.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
greenlet==3.0.3
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.3
Mako==1.3.2
MarkupSafe==2.1.5
mysqlclient==2.2.4
packaging==23.2
python-dotenv==1.0.1
setuptools==69.1.1
SQLAlchemy==2.0.27
SQLAlchemy-Utils==0.41.1
typing_extensions==4.9.0
Werkzeug==3.0.1
WTForms==3.1.2